feat(categoria): implementa controller e views com partials

+++ ControleFinanceiro.Domain/Interfaces/ICategoriaRepository.cs
@@ -5,9 +5,9 @@ namespace ControleFinanceiro.Domain.Repositories;
 public interface ICategoriaRepository : IDisposable
 {
     List<Categoria> Listar();
-    Categoria ListarPorId(int id);
+    Categoria ListarPorId(long id);
     List<Categoria> ListarPorNome(string categoria);
     void Criar(Categoria categoria);
     void Atualizar(Categoria categoria);
-    void Deletar(int id);
+    void Deletar(long id);
 }
